requirements.txt
astroid==2.3.3
colorama==0.4.3
cycler==0.10.0
dnspython==1.16.0
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
mysql-connector-python==8.0.19
numpy==1.18.4
protobuf==3.6.1
pylint==2.4.4
pyparsing==2.4.7
pyperclip==1.8.0
python-dateutil==2.8.1
six==1.14.0
wrapt==1.11.2
